moving some functions from geo utils to separate preprocessing python and other things
